Make custom hook for retrieving item list from database
